moved type.h and applied changes for all codes
