+++ include/draw/draw.h
@@ -4,7 +4,7 @@
 #include <cstddef>
 #include <opencv2/opencv.hpp>
 
-#include "type.h"
+#include "pcdet-infer-cpu/common/type.h"
 
 inline cv::Point2f rotatePoint(const cv::Point2f &point, const float angle) {
   const float x = point.x * std::cos(angle) - point.y * std::sin(angle);
